+++ material/src/plugins/contentRenderer/components.ts
@@ -1,0 +1,3 @@
+export * from './virtualScrollTable/virtualScrollTableContentRenderer.component';
+export * from './virtualScrollTable/header/virtualScrollTableHeaderContentRenderer.component';
+export * from './virtualScrollTable/body/virtualScrollTableBodyContentRenderer.component';