+++ material/src/plugins/contentRenderer/index.ts
@@ -1,0 +1,1 @@
+export * from './virtualScrollTable/virtualScrollTableContentRenderer.interface';